Extracted FilterableTableStorageRepository to its own file
IFilterableTableStorageRepository inherits from ITableStorageRepository
