+++ src/ForEvolve.Azure/Storage/Table/IFilterableTableStorageRepository.cs
@@ -5,8 +5,8 @@ using System.Threading.Tasks;
 
 namespace ForEvolve.Azure.Storage.Table
 {
-    public interface IFilterableTableStorageRepository<TModel>
-         where TModel : ITableEntity, new()
+    public interface IFilterableTableStorageRepository<TModel> : ITableStorageRepository<TModel>
+        where TModel : class, ITableEntity, new()
     {
         Task<IEnumerable<TModel>> ReadAsync(Func<TableQuery<TModel>, TableQuery<TModel>> filter);
     }
